+++ docs/src/examples/query/events/query-on-next/page.js
@@ -1,0 +1,13 @@
+import { $ } from '@semantic-ui/query';
+
+const addLog = (text) => {
+  $('.log').append(`<div>${text}</div>`);
+};
+
+$('.click').on('click', async function() {
+  addLog('Clicked');
+
+  // async style .one()
+  await $(this).onNext('mouseleave');
+  addLog('Left');
+});